Atualização: adicionada validação nos metodos de deletar e login

+++ src/main/java/medplanner/controller/UsuarioController.java
@@ -7,11 +7,16 @@ import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
@@ -82,22 +87,18 @@ public class UsuarioController {
         errors.addAll(result.getFieldErrors().stream().map(FieldError::getDefaultMessage).collect(Collectors.toList()));
     }
         
-        // Validar CPF
         if (!cpfValidator.isValid(usuario.getCpf(), null)) {
             errors.add("CPF inválido");
         }
     
-        // Verificar se o usuário já existe
         if (usuarioRepository.findByUsername(usuario.getUsername()) != null) {
             errors.add("Usuário já existe.");
         }
     
-        // Verificar se o CPF já existe
         if (usuarioRepository.findByCPF(usuario.getCpf()) != null) {
             errors.add("CPF já existe.");
         }
     
-        // Se houver erros, retornar a lista de erros
         if (!errors.isEmpty()) {
             Map<String, List<String>> errorResponse = new HashMap<>();
             errorResponse.put("errors", errors);
@@ -114,37 +115,44 @@ public class UsuarioController {
         }
     }
 
-
     @DeleteMapping("/deletar/{id}")
-    public ResponseEntity<String> deletarUsuario(@PathVariable Long id) {
-        Usuario usuario = usuarioRepository.findById(id).orElse(null);
-
-        if (usuario != null) {
-            usuarioRepository.delete(usuario);
-            return ResponseEntity.ok("Usuário deletado com sucesso");
-        } else {
-            return ResponseEntity.notFound().build();
-        }
+    public ResponseEntity<String> deletarUsuario(@PathVariable Long id, @AuthenticationPrincipal UserDetails userDetails) {
+    if (!userDetails.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMINISTRADOR"))) {
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Apenas usuários com cargo de ADMINISTRADOR podem excluir registros.");
     }
+
+    Usuario usuario = usuarioRepository.findById(id).orElse(null);
+
+    if (usuario != null) {
+        usuarioRepository.delete(usuario);
+        return ResponseEntity.ok("Usuário deletado com sucesso");
+    } else {
+        return ResponseEntity.notFound().build();
+    }
+}
 
     @PostMapping("/login")
     public ResponseEntity login(@RequestBody Usuario usuario) {
-        try {
-            var usernamePassword = new UsernamePasswordAuthenticationToken(usuario.getUsername(),
-                    usuario.getPassword());
-            var auth = this.authenticationManager.authenticate(usernamePassword);
-
-            Usuario usuarioAutenticado = (Usuario) auth.getPrincipal();
-
-            var token = tokenService.generateToken(usuarioAutenticado);
-
-            Map<String, String> response = new HashMap<>();
-            response.put("token", token);
-            response.put("usuario", usuarioAutenticado.getNome());
-
-            return ResponseEntity.ok(response);
-        } catch (Exception e) {
-            return ResponseEntity.badRequest().body(e.getMessage());
-        }
+    if (StringUtils.isEmpty(usuario.getUsername()) || StringUtils.isEmpty(usuario.getPassword())) {
+        return ResponseEntity.badRequest().body("Nome de usuário e senha são obrigatórios.");
     }
+
+    try {
+        var usernamePassword = new UsernamePasswordAuthenticationToken(usuario.getUsername(),
+                usuario.getPassword());
+        var auth = this.authenticationManager.authenticate(usernamePassword);
+
+        Usuario usuarioAutenticado = (Usuario) auth.getPrincipal();
+
+        var token = tokenService.generateToken(usuarioAutenticado);
+
+        Map<String, String> response = new HashMap<>();
+        response.put("token", token);
+        response.put("usuario", usuarioAutenticado.getNome());
+
+        return ResponseEntity.ok(response);
+    } catch (Exception e) {
+        return ResponseEntity.badRequest().body(e.getMessage());
+    }
+}
 }
